requirements.txt
altair
numpy
pandas
pydeck
streamlit
langchain
pinecone-client
boto3
toml
openai